Full AuthLogs functionality implemented
the user can login, the time and date will be saved in the AuthLogs by creating a new AuthLog Id. if the user wish to sign out the Signout button will update the AuthLog table and add the current time in the Logout_time. the Admin will be see when the user logged on and sign out